pyproject.toml
[project]
name = "wardrobe"
version = "0.1.0"
description = "Generative AI mobile app that recommends outfits based on your collection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "pydantic[email]>=2.12.5",
]
[project.optional-dependencies]
tryon = ["pillow>=10.0.0"]
diffusion-tryon = ["gradio_client>=0.15.0"]

[project.scripts]
wardrobe = "wardrobe.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wardrobe"]
